backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
pgvector>=0.3.6
python-dotenv>=1.0.1
python-multipart>=0.0.20

openai>=1.61.0
langchain>=0.3.17
langchain-community>=0.3.16
langchain-openai>=0.3.3
langchain-postgres>=0.0.13

pymupdf4llm>=0.0.17
rank-bm25>=0.2.2
sentence-transformers>=3.4.1
numpy>=2.2.2
pydantic>=2.10.6

camelot-py[cv]>=1.0.0
tabula-py>=2.10.0
